Added built-in function type and converted the print statement into the print built-in.
Value was moved from vm.h to value.h through some creative typedef'ing
The Builtin enum is not bounds-checked because it's not an enum, just a typedef of size_t. This is due to a conflict where C needed Value in builtin.h, but it also needed Builtin in value.h. However, it doesn't know beforehand what integer type it will use for Builtin, I guess, so there's a circular dependency.